requirements.txt
lightning
matplotlib
numpy
pandas
PyYAML
scikit_learn
seaborn
tensorboard
torch
